[feat] 물체 이동, 방향 회전 구현
