+++ Makefile
@@ -1,9 +1,24 @@
 NAME = miniRT
-SRCS = main.c \
+CFILES = main.c \
 		draw.c \
-		vector_util1.c
+		vector_util1.c \
+		vector_util2.c \
+		vector_util3.c \
+		parse_file.c \
+		parse_util.c \
+		parse_major_object.c \
+		parse_minor_object.c \
+		init.c \
+		hit_check.c \
+		key_util.c \
+		mouse_click.c \
+		key_move.c \
+		key_rotate_and_util.c
 
-HDRS = headers
+SRC_DIR = ./srcs
+HDRS = ./headers
+
+SRCS = ${addprefix ${SRC_DIR}/, ${CFILES}}
 
 OBJS = $(SRCS:.c=.o)
 CFLAGS = -Wall -Wextra -Werror -g
