+++ headers/key_code.h
@@ -1,0 +1,27 @@
+#ifndef KEY_CODE_H
+# define KEY_CODE_H
+
+/* key for move */
+# define KEY_A 0
+# define KEY_S 1
+# define KEY_D 2
+# define KEY_Q 12
+# define KEY_W 13
+# define KEY_E 14
+
+# define KEY_DELETE 51
+# define KEY_ESC 53
+# define KEY_L 37
+# define KEY_C 8
+
+/* key for rotate */
+# define KEY_Z 6
+# define KEY_X 7
+# define KEY_LEFT 123
+# define KEY_RIGHT 124
+# define KEY_UP 126
+# define KEY_DOWN 125
+
+# define LEFT_CLICK 1
+
+#endif
