+++ headers/structs.h
@@ -1,33 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   structs.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nheo <nheo@student.42seoul.kr>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/11/21 12:40:57 by nheo              #+#    #+#             */
+/*   Updated: 2022/11/21 16:35:52 by nheo             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef STRUCTS_H
 # define STRUCTS_H
 
-# include "./libft/headers/libft.h"
-# include "./mlx/mlx.h"
-#include <stddef.h>
-#include <sys/fcntl.h>
-#include <unistd.h>
-#include <fcntl.h>
-#include <stdlib.h>
-#include <math.h>
-
-# define WIDTH 1200
-# define HEIGHT 900
-# define TRUE 1
-# define FALSE 0
-
-typedef enum e_obj
-{
-	SPHERE,
-	PLANE,
-	CYLINDER
-}	t_object_type;
-
-typedef enum e_info_type
-{
-	AMB	= 0x0001,
-	CAM	= 0x0002,
-	LIGHT	= 0x0004,
-}	t_info_type;
+# include "libft.h"
 
 typedef struct s_pt
 {
@@ -46,13 +32,13 @@ typedef struct s_cam
 {
 	struct s_pt	pos;
 	struct s_pt	dir;
-	double		focal_len;
-	double		v_h; // 뷰포트 세로길이
-	double		v_w; // 뷰포트 가로길이
 	struct s_pt	vp_ll; // 뷰포트 왼쪽 아래 꼭지점
 	struct s_pt	hor; // 뷰포트 가로 방향 벡터
 	struct s_pt	ver; // 뷰포트 세로 방향 벡터
+	double		v_h; // 뷰포트 세로길이
+	double		v_w; // 뷰포트 가로길이
 	double		fov;
+	double		focal_len;
 }	t_cam;
 
 typedef struct s_light
@@ -97,10 +83,10 @@ typedef struct s_hit_check
 {
 	t_pt		pos;	// 광선과 object가 만나는 지점
 	t_pt		n_vec;	// 법선 벡터
+	t_pt		albedo;	// object의 색깔
 	double		t;		// 광선과 object가 만나는 지점까지의 거리
 	double		t_min;	// 광선과 object가 만나는 지점까지의 최소 거리 (화면에 나타날 수 있는 최소 거리)
 	double		t_max;	// 광선과 object가 만나는 지점까지의 최대 거리 (화면에 나타날 수 있는 최대 거리)
-	t_pt		albedo;	// object의 색깔
 	int			is_surface;	// object의 surface에 있는지 체크
 }	t_hit_check;
 
@@ -114,15 +100,15 @@ typedef struct s_ray
 typedef struct s_pong
 {
 	t_pt		lig_dir;
-	double		len;
-	double		kd;
-	double		ks;
-	double		ksn;
-	t_pt		dif;	// 난반사
-	double		spec;	// 입사각, 반사각 사이의 각도
 	t_pt		specular;	// 반사광
 	t_pt		view_dir;
 	t_pt		ref_dir;
+	t_pt		dif;	// 난반사
+	double		lig_len;
+	double		kd;
+	double		ks;
+	double		ksn;
+	double		spec;	// 입사각, 반사각 사이의 각도
 }	t_pong;
 
 typedef struct s_info
@@ -139,21 +125,9 @@ typedef struct s_info
 	struct s_ambient	amb;
 	struct s_cam		cam;
 	struct s_light		light;
+	void				*clicked;
+	int					clicked_type;
 	t_list				*objs;
 }	t_info;
-
-double	vlength2(t_pt vec);
-double	vlength(t_pt vec);
-t_pt	vunit(t_pt vec);
-t_pt	vadd(t_pt vec1, t_pt vec2);
-t_pt	vsub(t_pt vec1, t_pt vec2);
-t_pt	vmult(t_pt vec, double num);
-t_pt	vmult_vec(t_pt vec1, t_pt vec2);
-t_pt	vdiv(t_pt vec, double t);
-double	vdot(t_pt vec1, t_pt vec2);
-t_pt	vcross(t_pt vec1, t_pt vec2);
-t_pt	vreflect(t_pt vec, t_pt n_vec);
-t_pt	vmin(t_pt vec1, t_pt vec2);
-void	exceve(t_info *info);
 
 #endif
