+++ tmp.rt
@@ -1,7 +1,7 @@
 A         0.15         255,255,255
 C         0,0,0                  0,0,-1                  90
-L         0,-10,0                           0.8          120,120,200
+L         0,-10,0                           0.8          255,0,255
 
 sp 3,0,-10 2 0,0,255
 sp 3,0,-5 2 255,0,0
-pl         0,5,0                           0,1,0         255,255,0
+pl         0,5,0                           0,1,0         255,0,255
